api/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Validation & Settings
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Redis
redis==5.1.1

# Scheduler
apscheduler==3.10.4

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
google-auth==2.38.0
requests==2.32.3
eth-account==0.13.4
PyNaCl==1.5.0
base58==2.1.1

# AI Integration
anthropic==0.34.0
groq==0.11.0

# Object Storage (S3-compatible)
boto3==1.35.0
Pillow==10.4.0

# HTTP client (Privy JWKS, Base RPC)
httpx==0.27.2

# Dev & Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Linting
ruff==0.6.8
